# Re-factor machine-level optimisation decision logic for x25519