### Issues:

CryptoAlg-1602

### Description of changes: 

PR chain:
* [DONE] PR1: factor out C-level implementation
* PR2: refator backend implementation selection business logic to make it more sane
* PR3: hook-up s2n-bignum x25519 backend to selection business logic
* PR4: retire existing x25519 arm implementation

After factoring out the C-implementation in https://github.com/aws/aws-lc/commit/c5e2fb892b22c88b6cdd992a4bd52d93590d263b, now we factor out the control flow decision logic that selects the backend algorithm. This makes it simpler to add new backends.

The logic is modified as follows:
**before**
```
x25519_scalar_mult() {
#if defined(BORINGSSL_X25519_NEON)
  x25519_neon();
#endif

  x25519_nohw();
}

x25519() {
  x25519_scalar_mult();
  // zero key check
}

X25519_public_from_private() {
#if defined(BORINGSSL_X25519_NEON)
  X25519_public_from_private_neon();
#endif

  X25519_public_from_private_nohw();
}
```

**after**
```
x25519() {
  if (curve25519_asm_capable) {
    x25519_neon();
  } else {
    x25519_nohw();
  }
  // zero key check
}

X25519_public_from_private() {
  if (curve25519_asm_capable) {
    X25519_public_from_private_neon();
  } else {
    X25519_public_from_private_nohw();
  }
}
```

### Call-outs:

I moved the function `ED25519_keypair_from_seed()` because currently it's called by a function that is placed before it. Just a bit a code movement.

Also move some code from the internal header file to `curve25519.c`, since it doesn't need to be shared.

### Testing:

Sniff test for performance

#### Arm

**before**

```
$ ./tool/bssl speed -filter 25519
Did 18000 Ed25519 key generation operations in 1012849us (17771.7 ops/sec)
Did 18000 Ed25519 signing operations in 1025782us (17547.6 ops/sec)
Did 6039 Ed25519 verify operations in 1090322us (5538.7 ops/sec)
Did 19000 Curve25519 base-point multiplication operations in 1051713us (18065.8 ops/sec)
Did 7139 Curve25519 arbitrary point multiplication operations in 1098980us (6496.0 ops/sec)
Did 4780 ECDH X25519 operations in 1001049us (4775.0 ops/sec)
```

**after**

```
$ ./tool/bssl speed -filter 25519
Did 18000 Ed25519 key generation operations in 1011521us (17795.0 ops/sec)
Did 18000 Ed25519 signing operations in 1028718us (17497.5 ops/sec)
Did 6072 Ed25519 verify operations in 1090067us (5570.3 ops/sec)
Did 19000 Curve25519 base-point multiplication operations in 1052668us (18049.4 ops/sec)
Did 7139 Curve25519 arbitrary point multiplication operations in 1098808us (6497.0 ops/sec)
Did 5236 ECDH X25519 operations in 1096061us (4777.1 ops/sec)
```

#### Intel

**before**

```
$ ./tool/bssl speed -filter 25519
Did 56000 Ed25519 key generation operations in 1017600us (55031.4 ops/sec)
Did 55000 Ed25519 signing operations in 1014982us (54188.2 ops/sec)
Did 21000 Ed25519 verify operations in 1031452us (20359.6 ops/sec)
Did 58000 Curve25519 base-point multiplication operations in 1011957us (57314.7 ops/sec)
Did 25000 Curve25519 arbitrary point multiplication operations in 1022885us (24440.7 ops/sec)
Did 18000 ECDH X25519 operations in 1057809us (17016.3 ops/sec)
```


**after**

```
$ ./tool/bssl speed -filter 25519
Did 55000 Ed25519 key generation operations in 1016495us (54107.5 ops/sec)
Did 54000 Ed25519 signing operations in 1013897us (53259.8 ops/sec)
Did 21000 Ed25519 verify operations in 1022909us (20529.7 ops/sec)
Did 57000 Curve25519 base-point multiplication operations in 1008834us (56500.9 ops/sec)
Did 25000 Curve25519 arbitrary point multiplication operations in 1040498us (24027.0 ops/sec)
Did 17000 ECDH X25519 operations in 1000077us (16998.7 ops/sec)
```


By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license and 
the ISC license.